Do the truncating in ClipListView's FrameResized() instead of the ClipListItem's DrawItem(). Remove ClipListView's Draw() and do SetViewColor in the MainWindow. Use flag B_FULL_UPDATE_ON_RESIZE for ClipListView.

+++ ClipdingerSettings.h
@@ -30,7 +30,7 @@ public:
 		void		SetLimit(int32 limit) { fLimit = limit; }
 		void		SetFade(int32 fade) { fFade = fade; }
 		void		SetFadeDelay(int32 delay) { fFadeDelay = delay; }
-		void		SetFadeStep(float step) { fFadeStep = step; }
+		void		SetFadeStep(int32 step) { fFadeStep = step; }
 		void		SetWindowPosition(BRect where) { fPosition = where; }
 private:
 		int32		fLimit;
